Move build sets into template metadata package
