+++ workspaces/apps/packages/goldstack-home/package.json
@@ -38,7 +38,7 @@
     "@goldstack/infra-aws": "0.3.35",
     "@goldstack/project-config": "0.1.0",
     "@goldstack/project-repository": "0.1.0",
-    "@goldstack/project-template-data": "0.1.0",
+    "@goldstack/template-metadata": "0.1.0",
     "@goldstack/utils-package": "0.3.30",
     "@goldstack/utils-project": "0.3.26",
     "@goldstack/utils-track": "0.1.12",
