+++ workspaces/apps/packages/goldstack-home/src/components/template/imageUtil.ts
@@ -14,7 +14,7 @@ import LambdaIcon from './../../icons/lambda.svg';
 import SESIcon from './../../icons/ses2.svg';
 import APIGateway from './../../icons/aws-api-gateway.svg';
 
-import { TemplateFeatureImage } from '@goldstack/project-template-data';
+import { TemplateFeatureImage } from '@goldstack/template-metadata';
 
 export const resolveImage = (image: TemplateFeatureImage): any => {
   if (image.indexOf('http') === 0) {
