+++ workspaces/apps/packages/template-management-cli/src/scheduleAllDeploySets.ts
@@ -1,5 +1,5 @@
 import { start } from './image';
-import { getAllBuildSets } from './deploySets/deploySets';
+import { getAllBuildSets } from '@goldstack/template-metadata';
 
 export const scheduleAllDeploySets = async (argv: any): Promise<void> => {
   const sets = await getAllBuildSets();
