--- workspaces/templates/packages/project-template-data/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "include": [
-    "./src/**/*"
-  ],
-  "compilerOptions": {
-    "outDir": "dist",
-    "baseUrl": "."
-  },
-  "references": []
-}
