+++ workspaces/templates/packages/template-metadata/.npmignore
