+++ workspaces/templates/packages/template-metadata/jest.config.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const base = require('./../../jest.config');
+const base = require('../../jest.config');
 
 module.exports = {
   ...base,
