+++ workspaces/templates/packages/template-metadata/src/deploySets/nextjsBootstrap.ts
