+++ workspaces/templates/packages/template-metadata/src/deploySets/staticWebsite.ts
