+++ workspaces/templates/packages/template-metadata/src/projectTemplateData.spec.ts
