+++ workspaces/templates/packages/template-metadata/src/projectTemplateData.ts
@@ -23,6 +23,7 @@ export { getNextJsTemplateData };
 export { getExpressTemplateData };
 export { getNextjsBootstrapTemplateData };
 export { getGoGinTemplateData };
+export { getAllBuildSets, getBuildSet } from './deploySets/deploySets';
 
 export const allTemplates = (): ProjectTemplateProps[] => {
   const templates = [
