+++ workspaces/templates/packages/template-metadata/src/projectTemplateTypes.ts
