+++ workspaces/templates/packages/template-metadata/src/templates/emailSentTemplateData.ts
