+++ workspaces/templates/packages/template-metadata/src/templates/goGinTemplateData.ts
