+++ workspaces/templates/packages/template-metadata/src/templates/lambdaApiTemplateData.ts
