+++ workspaces/templates/packages/template-metadata/src/templates/moduleData.ts
@@ -1,4 +1,4 @@
-import { PackageProps } from './../projectTemplateTypes';
+import { PackageProps } from '../projectTemplateTypes';
 
 const NextJsBootstrap: PackageProps = {
   title: 'Next.js + Bootstrap',
