+++ workspaces/templates/packages/template-metadata/src/templates/nextjsBootstrapTemplateData.ts
