+++ workspaces/templates/packages/template-metadata/src/templates/nextjsTemplateData.ts
